refactor: 백엔드 게시글 상세 페이지에서 ArticleLike를 가져오는 과정에서 성능 개선

+++ backend/rush/src/main/java/rush/rush/dto/ArticleResponse.java
@@ -18,5 +18,5 @@ public class ArticleResponse {
     private Double longitude;
     private AuthorResponse author;
     private Timestamp createDate;
-    private Long totalLikes;
+    private int totalLikes;
 }
